+++ src/batlock666/example/functions.py
@@ -1,0 +1,14 @@
+# NOQA: D100
+
+def hello(name=None):
+    """Return a hello message.
+    """
+    if name is None:
+        # No name was given
+        return "Hello, World!"
+    else:
+        # Name was given
+        return "Hello, %s!" % name
+
+
+__all__ = ["hello"]