+++ CHANGELOG.rst
@@ -5,4 +5,4 @@ Changelog
 0.1.0 (unreleased)
 ------------------
 
-- Nothing here yet.
+- Add function ``hello``.